requirements.txt
dash==2.0.0
dash_bootstrap_components==1.0.3
celery==5.2.3
numba==0.53.1
pandas==1.2.3
pytz==2021.3
Flask==2.0.2
pydantic==1.8.2
win10toast==0.9
numpy==1.20.2
Jinja2==3.0.1
requests==2.26.0
plotly==5.3.1
uvicorn==0.17.5
beautifulsoup4==4.10.0
fastapi==0.74.1
python-dotenv==0.19.2
websocket_client==1.3.1
